Fix authorize endpoint compliance - remove scope, make state optional
The scope parameter is optional in the Authorization Request (Section
4.1.1 and 4.2.1 of [RFC 6749] and Section 3.3.2.1 of [OIDC 1.0 Core]).

Authorization Endpoint does not include the scope parameter in the
response when the response type is "code" (Section 4.1.2 of [RFC 6749]
and 3.1.2.5 of [OIDC 1.0 Core]).

The state parameter is recommended, but not mandatory, according to
(Section 4.1.1 of [RFC 6749] and Section 3.3.2.1 of [OIDC 1.0 Core]).

[RFC 6749]: https://datatracker.ietf.org/doc/html/rfc6749
[OIDC 1.0 Core]: https://openid.net/specs/openid-connect-core-1_0.html